started removing extra camera state
