+++ packages/space-opera/src/components/camera_settings/camera_settings.ts
@@ -30,17 +30,14 @@ import {customElement, html, internalProperty, property, query} from 'lit-elemen
 import {reduxStore} from '../../space_opera_base.js';
 import {cameraSettingsStyles} from '../../styles.css.js';
 import {ModelViewerConfig, State} from '../../types.js';
-import {dispatchAutoRotate, getConfig} from '../config/reducer.js';
+import {dispatchAutoRotate, dispatchCameraTarget, dispatchRadiusLimits, dispatchSaveCameraOrbit, getConfig} from '../config/reducer.js';
+import {Limits, SphericalPositionDeg, Vector3D} from '../config/types.js';
 import {ConnectedLitElement} from '../connected_lit_element/connected_lit_element.js';
 import {getCameraState, getModelViewer} from '../model_viewer_preview/reducer.js';
 import {CheckboxElement} from '../shared/checkbox/checkbox.js';
 import {DraggableInput} from '../shared/draggable_input/draggable_input.js';
 import {styles as draggableInputRowStyles} from '../shared/draggable_input/draggable_input_row.css.js';
 import {checkFinite} from '../utils/reducer_utils.js';
-
-import {Camera, INITIAL_CAMERA} from './camera_state.js';
-import {dispatchCameraTarget, dispatchRadiusLimits, dispatchSaveCameraOrbit, getCamera} from './reducer.js';
-import {Limits, SphericalPositionDeg, Vector3D} from './types.js';
 
 @customElement('me-camera-orbit-editor')
 class CameraOrbitEditor extends ConnectedLitElement {
@@ -153,9 +150,6 @@ export class CameraSettings extends ConnectedLitElement {
   static styles = cameraSettingsStyles;
 
   @internalProperty() config: ModelViewerConfig = {};
-  @internalProperty() camera: Camera = INITIAL_CAMERA;
-  @internalProperty() initialCamera: Camera = INITIAL_CAMERA;
-  @internalProperty() cameraOutOfBounds: boolean = false;
 
   @query('me-camera-orbit-editor') cameraOrbitEditor?: CameraOrbitEditor;
   @query('me-checkbox#auto-rotate') autoRotateCheckbox!: CheckboxElement;
@@ -170,31 +164,12 @@ export class CameraSettings extends ConnectedLitElement {
 
   stateChanged(state: State) {
     this.config = getConfig(state);
-    this.camera = getCamera(state);
-    this.cameraOutOfBounds = this.outOfBounds();
   }
 
   orbitValueBound(limits: Limits|undefined, val: string|number) {
     if ((limits !== undefined) &&
         ((limits.max !== 'auto' && val > limits.max) ||
          (limits.min !== 'auto' && val < limits.min))) {
-      return true;
-    }
-    return false;
-  }
-
-  outOfBounds() {
-    const snippet = this.camera;
-    if (snippet.orbit === undefined) {
-      return false;
-    }
-    if (this.orbitValueBound(snippet.pitchLimitsDeg, snippet.orbit?.phiDeg)) {
-      return true;
-    } else if (this.orbitValueBound(
-                   snippet.yawLimitsDeg, snippet.orbit?.thetaDeg)) {
-      return true;
-    } else if (this.orbitValueBound(
-                   snippet.radiusLimits, snippet.orbit?.radius)) {
       return true;
     }
     return false;
@@ -237,7 +212,6 @@ export class CameraSettings extends ConnectedLitElement {
   }
 
   render() {
-    const initalError = this.cameraOutOfBounds ? 'initialError' : ''
     return html`
     <me-expandable-tab tabName="Camera Setup" .open=${true}>
       <div slot="content">
@@ -245,11 +219,11 @@ export class CameraSettings extends ConnectedLitElement {
           ?checked="${!!this.config.autoRotate}"
           @change=${this.onAutoRotateChange}>
         </me-checkbox>
-        <div class="${initalError}">
+        <div>
           <div style="font-size: 14px; font-weight: 500; margin-top: 10px">Initial Camera Position:</div>
           <me-camera-orbit-editor
             @change=${this.onCameraOrbitEditorChange}
-            .orbit=${this.camera.orbit ?? this.initialCamera.orbit}>
+            .orbit=${}>
           </me-camera-orbit-editor>
           <div style="justify-content: space-between; width: 100%; display: flex;">
             <mwc-button
@@ -264,12 +238,6 @@ export class CameraSettings extends ConnectedLitElement {
             <mwc-icon-button class="RevertButton" style="align-self: center; margin-top: 10px;" id="revert-metallic-roughness-texture" icon="undo"
             title="Reset initial camera" @click=${this.resetInitialCamera}>
             </mwc-icon-button>
-          </div>
-          ${
-        this.cameraOutOfBounds ?
-            html`<div class="error">Your initial camera is outside the bounds of your limits. Set your initial camera again.</div>` :
-            html``}
-        </div>
         <div style="font-size: 14px; font-weight: 500; margin-top: 20px">Target Point:</div>
         <me-camera-target-input .change=${this.onCameraTargetChange}>
         </me-camera-target-input>
