+++ packages/space-opera/src/components/camera_settings/components/limits_base.ts
@@ -23,10 +23,10 @@ import '../../shared/checkbox/checkbox.js';
 
 import {html, internalProperty, query} from 'lit-element';
 
+import {Limits} from '../../config/types.js';
 import {ConnectedLitElement} from '../../connected_lit_element/connected_lit_element.js';
 import {getModelViewer} from '../../model_viewer_preview/reducer.js';
 import {SliderWithInputElement} from '../../shared/slider_with_input/slider_with_input.js';
-import {Limits} from '../types.js';
 
 import {styles} from './limits_base.css.js';
