+++ packages/space-opera/src/components/camera_settings/components/yaw_limits.ts
@@ -19,9 +19,9 @@ import {customElement, internalProperty} from 'lit-element';
 
 import {reduxStore} from '../../../space_opera_base.js';
 import {State} from '../../../types.js';
+import {Limits} from '../../config/types.js';
 import {getCameraState, getModelViewer} from '../../model_viewer_preview/reducer.js';
-import {dispatchYawLimits, getCamera, getIsDirtyCamera} from '../reducer.js';
-import {Limits} from '../types.js';
+import {dispatchYawLimits, getIsDirtyCamera} from '../reducer.js';
 
 import {LimitsBase} from './limits_base.js';
