+++ packages/space-opera/src/components/camera_settings/reducer.ts
@@ -16,11 +16,6 @@
  */
 
 import {Action, State} from '../../types.js';
-import {INITIAL_CAMERA} from './camera_state.js';
-
-import {Camera} from './camera_state.js';
-import {SphericalPositionDeg, Vector3D} from './types.js';
-import {Limits} from './types.js';
 
 // DIRTY CAMERA//////////////
 
@@ -36,113 +31,6 @@ export function isDirtyCameraReducer(
   switch (action.type) {
     case IS_DIRTY_CAMERA:
       return !state;
-    default:
-      return state;
-  }
-}
-
-// CAMERA //////////////
-
-/** Dispatch change to maximum pitch */
-const SET_CAMERA_YAW_LIMITS = 'SET_CAMERA_YAW_LIMITS';
-export function dispatchYawLimits(yawLimitsDeg?: Limits) {
-  if (!yawLimitsDeg) {
-    throw new Error('No limits given');
-  }
-  return {type: SET_CAMERA_YAW_LIMITS, payload: yawLimitsDeg};
-}
-
-/** Dispatch change to radius limits */
-const SET_CAMERA_RADIUS_LIMITS = 'SET_CAMERA_RADIUS_LIMITS';
-export function dispatchRadiusLimits(radiusLimits?: Limits) {
-  return {type: SET_CAMERA_RADIUS_LIMITS, payload: radiusLimits};
-}
-
-/** Dispatch change to maximum pitch */
-const SET_CAMERA_PITCH_LIMITS = 'SET_CAMERA_PITCH_LIMITS';
-export function dispatchPitchLimits(pitchLimitsDeg?: Limits) {
-  return {type: SET_CAMERA_PITCH_LIMITS, payload: pitchLimitsDeg};
-}
-
-/** Dispatch change to maximum FOV */
-const SET_CAMERA_FOV_LIMITS = 'SET_CAMERA_FOV_LIMITS';
-export function dispatchFovLimits(fovLimitsDeg?: Limits) {
-  return {type: SET_CAMERA_FOV_LIMITS, payload: fovLimitsDeg};
-}
-
-const SET_MIN_ZOOM = 'SET_MIN_ZOOM';
-export function dispatchSetMinZoom(
-    fovDeg: number|string, radius: number|string) {
-  return {
-    type: SET_MIN_ZOOM, payload: {radius: radius, fov: fovDeg}
-  }
-}
-
-const SET_ZOOM_ENABLED = 'SET_ZOOM_ENABLED';
-export function dispatchZoomEnabled(isEnabled: boolean) {
-  return {
-    type: SET_ZOOM_ENABLED, payload: isEnabled
-  }
-}
-
-const SAVE_CAMERA_ORBIT = 'SAVE_CAMERA_ORBIT';
-export function dispatchSaveCameraOrbit(orbit: SphericalPositionDeg|undefined) {
-  return {type: SAVE_CAMERA_ORBIT, payload: orbit};
-}
-
-/** Event dispatcher for changes to camera-target. */
-const SET_CAMERA_TARGET = 'SET_CAMERA_TARGET';
-export function dispatchCameraTarget(target?: Vector3D) {
-  return {type: SET_CAMERA_TARGET, payload: target};
-}
-
-const SET_CAMERA = 'SET_CAMERA';
-export function dispatchSetCamera(camera: Camera) {
-  return {type: SET_CAMERA, payload: camera};
-}
-
-export const getCamera = (state: State) =>
-    state.entities.modelViewerSnippet.camera;
-
-export function cameraReducer(
-    state: Camera = INITIAL_CAMERA, action: Action): Camera {
-  switch (action.type) {
-    case SET_CAMERA:
-      return action.payload;
-    case SET_CAMERA_TARGET:
-      return {
-        ...state, target: action.payload
-      }
-    case SAVE_CAMERA_ORBIT:
-      return {...state, orbit: action.payload};
-    case SET_CAMERA_FOV_LIMITS:
-      return {
-        ...state, fovLimitsDeg: action.payload
-      }
-    case SET_CAMERA_PITCH_LIMITS:
-      return {
-        ...state, pitchLimitsDeg: action.payload
-      }
-    case SET_CAMERA_RADIUS_LIMITS:
-      return {
-        ...state, radiusLimits: action.payload
-      }
-    case SET_CAMERA_YAW_LIMITS:
-      return {
-        ...state, yawLimitsDeg: action.payload
-      }
-    case SET_MIN_ZOOM:
-      return {
-        ...state,
-            radiusLimits: {...state.radiusLimits!, min: action.payload.radius},
-            fovLimitsDeg: {...state.fovLimitsDeg!, min: action.payload.fov}
-      }
-    case SET_ZOOM_ENABLED:
-      return {
-        ...state,
-            radiusLimits: {...state.radiusLimits!, enabled: action.payload},
-            fovLimitsDeg: {...state.fovLimitsDeg!, enabled: action.payload}
-      }
     default:
       return state;
   }
